pyproject.toml
[tool.poetry]
name = "treasure-hunt"
version = "0.1.0"
description = ""
authors = ["shahroz  <shehroz_akbar@universalplus.technology>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "*"
amqp = "5.1.1"
celery = "5.4.0"
click = "8.1.7"
Django = "4.2.4"
django-celery-beat = "2.5.0"
django-celery-results = "2.5.1"
django-cors-headers = "4.2.0"
django-environ = "0.11.1"
django-filter = "23.2"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.0"
graphene-django = "3.1.5"
psycopg2-binary = "2.9.7"
redis = "5.0.0"
requests = "2.31.0"
django-oauth-toolkit = "^2.3.0"
social-auth-app-django = "^5.0.0"
drf-social-oauth2 = "^2.1.3"
django_redis = "^5.0.0"
pillow = "^10.0.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
